necChanges

+++ ft_printf.h
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   printf.h                                           :+:      :+:    :+:   */
+/*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iel-fagh <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: iel-fagh <iel-fagh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/07 07:58:33 by iel-fagh          #+#    #+#             */
-/*   Updated: 2023/12/09 12:52:43 by iel-fagh         ###   ########.fr       */
+/*   Updated: 2023/12/10 15:09:25 by iel-fagh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 int	ft_printf(const char *format, ...);
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
-int	ft_putnbr(long n);
+int	ft_putnbr(int n);
 int	ft_hex_lower(unsigned long n);
 int	ft_hex_upper(unsigned int n);
 int	ft_unsigned_int(unsigned int n);
